rename config.py to spalloc_config.py
